Remove multi-threaded loading for now
Documented *some* of ProfileManager
Removed ProfileManager.getFile in favor of ProfileManager.getLoadedFile
